pyproject.toml
[project]
name = "MAP"
version = "0.4.0"
description = "MAP: Drawing the map of research knowledge"
readme = "README.md"
requires-python = "<3.13,>=3.10"
dependencies = [
  "aiofiles==24.1.0",
  "aiolimiter==1.2.1",
  "annotated-types==0.7.0",
  "anyio==4.7.0",
  "anytree==2.12.1",
  "asttokens==2.4.1",
  "attrs==24.3.0",
  "autograd==1.7.0",
  "azure-common==1.1.28",
  "azure-core==1.32.0",
  "azure-cosmos==4.9.0",
  "azure-identity==1.19.0",
  "azure-search-documents==11.5.2",
  "azure-storage-blob==12.24.0",
  "beartype==0.18.5",
  "certifi>=2024.12.14",
  "cffi==1.17.1 ; platform_python_implementation != 'PyPy'",
  "charset-normalizer==3.4.0",
  "click==8.1.7",
  "colorama==0.4.6 ; platform_system == 'Windows'",
  "contourpy==1.3.1",
  "cryptography==44.0.0",
  "cycler==0.12.1",
  "datashaper==0.0.49",
  "deprecation==2.1.0",
  "devtools==0.12.2",
  "diskcache==5.6.3",
  "distro==1.9.0",
  "environs==11.2.1",
  "exceptiongroup==1.2.2 ; python_version < '3.11'",
  "executing==2.1.0",
  "fnllm==0.0.10",
  "fonttools==4.55.3",
  "future==1.0.0",
  "gensim==4.3.3",
  "graspologic-native==1.2.1",
  "graspologic==3.4.1",
  "h11==0.14.0",
  "httpcore==1.0.7",
  "httpx==0.28.1",
  "hyppo==0.4.0",
  "idna==3.10",
  "isodate==0.7.2",
  "jiter==0.8.2",
  "joblib==1.4.2",
  "json-repair==0.30.3",
  "jsonschema-specifications==2024.10.1",
  "jsonschema==4.23.0",
  "kiwisolver==1.4.7",
  "lancedb==0.17.0",
  "llvmlite==0.43.0",
  "markdown-it-py==3.0.0",
  "marshmallow==3.23.1",
  "matplotlib==3.10.0",
  "mdurl==0.1.2",
  "msal-extensions==1.2.0",
  "msal==1.31.1",
  "networkx==3.4.2",
  "nltk==3.9.1",
  "numba==0.60.0",
  "numpy==1.26.4",
  "openai==1.57.4",
  "overrides==7.7.0",
  "packaging==24.2",
  "pandas==2.2.3",
  "patsy==1.0.1",
  "pillow==11.0.0",
  "portalocker==2.10.1",
  "pot==0.9.5",
  "pyaml-env==1.2.1",
  "pyarrow==15.0.2",
  "pycparser==2.22 ; platform_python_implementation != 'PyPy'",
  "pydantic-core==2.27.1",
  "pydantic==2.10.3",
  "pygments==2.18.0",
  "pyjwt[crypto]==2.10.1",
  "pylance==0.20.0",
  "pynndescent==0.5.13",
  "pyparsing==3.2.0",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.0.1",
  "pytz==2024.2",
  "pywin32==308 ; platform_system == 'Windows'",
  "pyyaml==6.0.2",
  "referencing==0.35.1",
  "regex==2024.11.6",
  "requests==2.32.3",
  "rich==13.9.4",
  "rpds-py==0.22.3",
  "scikit-learn==1.6.0",
  "scipy==1.12.0",
  "seaborn==0.13.2",
  "shellingham==1.5.4",
  "six==1.17.0",
  "smart-open==7.0.5",
  "sniffio==1.3.1",
  "statsmodels==0.14.4",
  "tenacity==9.0.0",
  "threadpoolctl==3.5.0",
  "tiktoken==0.8.0",
  "tqdm==4.67.1",
  "typer==0.15.1",
  "typing-extensions>=4.12.2",
  "tzdata>=2024.2",
  "umap-learn==0.5.7",
  "urllib3>=2.2.3",
  "wrapt==1.17.0",
  "faiss-cpu>=1.11.0.post1",
]

[project.optional-dependencies]
dev = [
    "coverage<8.0.0,>=7.6.0",
    "ipykernel<7.0.0,>=6.29.4",
    "jupyter<2.0.0,>=1.1.1",
    "nbconvert<8.0.0,>=7.16.3",
    "poethepoet<1.0.0,>=0.28.0",
    "pyright<2.0.0,>=1.1.384",
    "pytest<9.0.0,>=8.3.2",
    "pytest-asyncio<1.0.0,>=0.24.0",
    "pytest-timeout<3.0.0,>=2.3.1",
    "ruff<1.0.0,>=0.6.9",
    "semversioner<3.0.0,>=2.0.3",
    "update-toml<1.0.0,>=0.2.1",
    "deptry<1.0.0,>=0.20.0",
    "mkdocs-material<10.0.0,>=9.5.39",
    "mkdocs-jupyter<1.0.0,>=0.25.0",
    "mkdocs-exclude-search<1.0.0,>=0.6.6",
    "mkdocs-typer<1.0.0,>=0.0.3",
    "pre-commit>=4.1.0"
]

preprocess = [
    "docling>=2.39.0",
    "selenium>=4.33.0",
    "beautifulsoup4>=4.13.3",
    "pypdf2>=3.0.1"
]

server = [
    "flask>=3.1.0",
    "flask_cors>=5.0.0",
    "flask_socketio>=5.5.1"
]

[tool.ruff]
include = ["*.py"]
line-length = 120
indent-width = 4
target-version = "py310"
src = ["graphrag", "util", "service"]

[tool.ruff.lint]
select = ["E", "F", "W", "B", "C"]
ignore = ["C408", "B024", "E203", "E402", "E501", "E721", "E731", "W293", "F401", "C901", "B006"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
force-sort-within-sections = false
combine-as-imports = false


[tool.pdm.build]
includes = ["graphrag"]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
